Account and Payment ver1

+++ AccountService/src/main/java/com/example/accountservice/dto/BalanceResponse.java
@@ -1,12 +1,15 @@
 package com.example.accountservice.dto;
 
+import java.math.BigInteger;
+import java.math.BigDecimal;
+
 public class BalanceResponse {
-    private Long userId;
-    private Long newBalance;
+    private BigInteger userId;
+    private BigDecimal newBalance;
 
-    public Long getUserId() { return userId; }
-    public void setUserId(Long userId) { this.userId = userId; }
+    public BigInteger getUserId() { return userId; }
+    public void setUserId(BigInteger userId) { this.userId = userId; }
 
-    public Long getNewBalance() { return newBalance; }
-    public void setNewBalance(Long newBalance) { this.newBalance = newBalance; }
+    public BigDecimal getNewBalance() { return newBalance; }
+    public void setNewBalance(BigDecimal newBalance) { this.newBalance = newBalance; }
 }
