+++ AccountService/src/main/java/com/example/accountservice/dto/TransactionRequest.java
@@ -3,15 +3,17 @@ package com.example.accountservice.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigInteger;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransactionRequest {
-	private Long userId;
+	private BigInteger userId;
 	private String mssv;
 	private String type;
-	private Long amount;
+	private BigDecimal amount;
 	private String description;
 	private String transactionId;
 }
