+++ AccountService/src/main/java/com/example/accountservice/dto/UnlockRequest.java
@@ -3,11 +3,12 @@ package com.example.accountservice.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigInteger;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class UnlockRequest {
-    private Long userId;
+    private BigInteger userId;
     private String lockKey;
 }
