+++ AccountService/src/main/java/com/example/accountservice/repository/AccountRepository.java
@@ -2,10 +2,11 @@ package com.example.accountservice.repository;
 
 import com.example.accountservice.model.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
-    Optional<Account> findByUserId(Long userId);
+    Optional<Account> findByUserId(BigInteger userId);
 }
