+++ PaymentService/src/main/java/com/example/paymentservice/client/AccountServiceClient.java
@@ -8,6 +8,8 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
+import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -23,7 +25,7 @@ public class AccountServiceClient {
         this.accountBaseUrl = accountBaseUrl;
     }
 
-    public Boolean lockUser(Long userId) {
+    public Boolean lockUser(BigInteger userId) {
         String url = accountBaseUrl + "/accounts/{userId}/lock";
         Map<String, Object> body = new HashMap<>();
         body.put("userId", userId);
@@ -38,7 +40,7 @@ public class AccountServiceClient {
         }
     }
 
-    public Boolean unlockUser(Long userId) {
+    public Boolean unlockUser(BigInteger userId) {
         String url = accountBaseUrl + "/accounts/{userId}/unlock";
         Map<String, Object> body = new HashMap<>();
         body.put("userId", userId);
@@ -47,21 +49,21 @@ public class AccountServiceClient {
         return res.getBody() != null && Boolean.TRUE.equals(res.getBody().getUnlocked());
     }
 
-    public Long getBalance(Long userId) {
+    public BigInteger getBalance(BigInteger userId) {
         String url = accountBaseUrl + "/accounts/{userId}/balance";
-        return restTemplate.getForObject(url, Long.class, userId);
+        return restTemplate.getForObject(url, BigInteger.class, userId);
     }
     
-    public Boolean getAccount(Long userId) {
+    public Boolean getAccount(BigInteger userId) {
         String url = accountBaseUrl + "/accounts/{userId}/balance";
         try {
-            restTemplate.getForObject(url, Long.class, userId);
+            restTemplate.getForObject(url, BigInteger.class, userId);
             return true; // Account exists
         } catch (Exception ex) {
             return false; // Account not found
         }
     }
-    public Boolean checkBalance(Long userId, Long amount) {
+    public Boolean checkBalance(BigInteger userId, BigDecimal amount) {
         String url = accountBaseUrl + "/accounts/checkBalance";
         Map<String, Object> requestBody = new HashMap<>();
         requestBody.put("userId", userId);
@@ -70,7 +72,7 @@ public class AccountServiceClient {
         return  res.getBody() != null && res.getBody();
     }
 
-    public Boolean updateBalance(Long userId, Long amount, String transactionId) {
+    public Boolean updateBalance(BigInteger userId, BigDecimal amount, String transactionId) {
         String url = accountBaseUrl + "/accounts/{userId}/balance";
         Map<String, Object> body = new HashMap<>();
         body.put("userId", userId);
@@ -86,7 +88,7 @@ public class AccountServiceClient {
         return res.getStatusCode().is2xxSuccessful() && res.getBody() != null && res.getBody().getNewBalance() != null;
     }
 
-    public String saveTransaction(Long userId, String transactionId, Long amount) {
+    public String saveTransaction(BigInteger userId, String transactionId, BigDecimal amount) {
         String url = accountBaseUrl + "/transactions";
         Map<String, Object> body = new HashMap<>();
         body.put("userId", userId);
@@ -102,13 +104,13 @@ public class AccountServiceClient {
     public static class LockResponse {
         private Boolean locked;
         private String lockKey;
-        private Long expiry;
+        private BigInteger expiry;
         public Boolean getLocked() { return locked; }
         public void setLocked(Boolean locked) { this.locked = locked; }
         public String getLockKey() { return lockKey; }
         public void setLockKey(String lockKey) { this.lockKey = lockKey; }
-        public Long getExpiry() { return expiry; }
-        public void setExpiry(Long expiry) { this.expiry = expiry; }
+        public BigInteger getExpiry() { return expiry; }
+        public void setExpiry(BigInteger expiry) { this.expiry = expiry; }
     }
 
     public static class UnlockResponse {
